Some restructuring to views and other misc files
