+++ src/components/menus/visualization-wizard/viz-components/customizations/LineColorControl.vue
@@ -23,7 +23,6 @@ watch(lineColor, (val) => {
       }
     });
   }
-  console.log(lineId.value, vwStore.visualizationCustomizationOptions)
 })
 
 onMounted(() => {
